+++ packages/server/src/services/setup.ts
@@ -3,7 +3,7 @@ import { AblyService } from './clientCommunication/ablyService';
 import 'dotenv/config';
 import { initializeServerDatabase } from './database';
 import { initializePubSub, pubSub } from './clientCommunication/pubsub';
-import globalData from '../../global.json';
+import globalData from '../../world_assets/global.json';
 import { initializeGameWorld } from './gameWorld/gameWorld';
 import { ServerWorldDescription } from './gameWorld/worldMetadata';
 import { initializeKnowledgeDB } from '@rt-potion/converse';
@@ -40,10 +40,8 @@ async function initializeAsync() {
   }
 
   logger.log(`loading world ${worldID}`);
-  const worldDataResponse = await fetch(
-    `https://potions.gg/world_assets/${worldID}/server/world_specific.json`
-  );
-  const worldSpecificData = await worldDataResponse.json();
+
+  const worldSpecificData = await import(`../../world_assets/${worldID}/world_specific.json`);
 
   try {
     await downloadData(supabase, worldID);
