+++ packages/server/world_assets/fire-world/world_specific.json
