--- packages/server/world_assets/test-world/client/world_specific.json
@@ -1,72 +1,0 @@
-{
-    "tiles": [
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,1,1,1,51,51,51,51,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,2,2,2,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,1,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,1,1,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,2,2,2,2,2,1,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,2,2,2,2,2,2,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,51,51,51,51,51,1,1,1,1,1,2,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,1,1,1,1,2,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,1,1,1,1,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,1,1,1,1,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,1,1,1,1,1,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,1,1,1,1,1,1,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,1,1,1,1,1,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,51,51,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,51,51,51,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,51,51,51,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,51,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,51,51,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,51,51,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,51,51,1,1,2,2,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,51,51,51,1,1,2,-1,-1],
-        [-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,51,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,51,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,51,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,51,51,51,51,51,51,51,51,1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,51,51,51,51,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,1,1,1,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1]
-    ],
-    "terrain_types": [
-        {
-            "name": "sand",
-            "id": 2,
-            "walkable": true
-        },
-        {
-            "name": "grass",
-            "id": 1,
-            "walkable": true
-        },
-        {
-            "name": "water",
-            "id": 51,
-            "walkable": false
-        }
-    ],
-"portals": []
-}
