+++ packages/server/world_assets/test-world/world_specific.json
