--- packages/server/world_assets/water-world/client/world_specific.json
@@ -1,66 +1,0 @@
-{
-    "tiles": [
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,1,1,1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,2,2,2,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,1,1,2,2,2,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,1,1,2,2,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,2,2,2,2,1,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,-1,1,2,2,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,2,2,2,2,2,2,2,-1,-1,-1],
-        [-1,-1,-1,-1,1,1,2,2,2,-1,-1,-1,-1,-1,2,2,2,2,2,2,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,1,1,1,1,1,2,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,1,1,2,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,1,1,1,1,2,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,1,2,2,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,1,1,1,1,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,1,1,1,1,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,1,1,2,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,2,2,-1,-1,-1,-1],
-        [-1,-1,1,1,2,2,2,2,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,2,2,-1,-1,-1],
-        [-1,-1,-1,1,1,2,2,2,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,-1,-1,-1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,2,-1,-1,-1],
-        [-1,-1,-1,1,1,2,2,-1,-1,-1,-1,-1,2,2,1,1,1,1,1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,1,1,2,2,2,-1,-1,-1,-1,-1,2,2,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,1,2,2,2,2,-1,-1,-1,-1,2,2,2,1,1,1,1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,2,-1,-1],
-        [-1,-1,1,2,2,2,2,-1,-1,-1,2,2,2,2,1,1,1,1,1,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,2,-1,-1],
-        [-1,-1,1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,2,2,-1],
-        [-1,-1,-1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,2,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,2,2,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,1,2,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,-1,-1,-1,-1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,1,1,1,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,1,-1,-1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,2,2,2,-1,-1,-1,-1,-1,1,1,1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,2,2,2,2,2,-1,-1,-1,-1,-1,-1,-1,1,1,1,1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1],
-        [-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1]
-    ],
-    "terrain_types": [
-        {
-            "name": "sand",
-            "id": 1,
-            "walkable": true
-        },
-        {
-            "name": "grass",
-            "id": 2,
-            "walkable": true
-        }
-    ],
-"portals": []
-}
