+++ packages/server/world_assets/water-world/world_specific.json
